refactor config elements out of strategies, refactor cli config, add wif key to config
