+++ dexbot/cli.py
@@ -93,7 +93,8 @@ def run(ctx):
             signal.signal(signal.SIGHUP, kill_workers)
             # TODO: reload config on SIGUSR1
             # signal.signal(signal.SIGUSR1, lambda x, y: worker.do_next_tick(worker.reread_config))
-        except ValueError:
+        except AttributeError:
+#        except ValueError:
             log.debug("Cannot set all signals -- not available on this platform")
         if ctx.obj['systemd']:
             try:
@@ -128,6 +129,7 @@ def runservice(ctx):
     click.echo("Starting dexbot daemon")
     os.system("systemctl --user start dexbot")
 
+    
 @main.command()
 @click.pass_context
 @configfile
